added models folder to utils
